% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hsci.R
\name{coloropt_na_value}
\alias{coloropt_na_value}
\title{extract the coloropt final neutral gray value for different palettes}
\usage{
coloropt_na_value(n, option = "normal")
}
\arguments{
\item{n}{the number of colors required from the palette (max 13 for normal, 7 for the other palettes)}

\item{option}{coloropt palette from normal, bright, dark, fancy and tarnish}
}
\description{
extract the coloropt final neutral gray value for different palettes
}
