% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hsci.R
\name{coloropt_pal}
\alias{coloropt_pal}
\title{coloropt palette}
\usage{
coloropt_pal(option = "normal")
}
\arguments{
\item{option}{coloropt palette from normal, bright, dark, fancy and tarnish}
}
\description{
coloropt palette
}
