% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hsci.R
\name{convert_palette_to_grayscale}
\alias{convert_palette_to_grayscale}
\title{convert a palette to grayscale}
\usage{
convert_palette_to_grayscale(palette)
}
\arguments{
\item{palette}{palette to convert}
}
\description{
convert a palette to grayscale
}
