% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghsci-package.R
\docType{package}
\name{gghsci-package}
\alias{gghsci}
\alias{gghsci-package}
\title{gghsci: HSCI ggplot2 theme}
\description{
ggplot2 theme used by the Human Sciences -- Computing Interaction research group at the University of Helsinki.
}
\keyword{internal}
