% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hsci.R
\name{save_plots}
\alias{save_plots}
\title{Save plots to PNG + SVG files}
\usage{
save_plots(dir, basename, plot, width = 6, height = 5, units = "in", dpi = 300)
}
\arguments{
\item{dir}{directory to save the plots to}

\item{basename}{the base name for the plot files}

\item{plot}{the plot to save}

\item{width}{plot width}

\item{height}{plot height}

\item{units}{the units in which width & height are measured}

\item{dpi}{plot resolution (dots per inch)}
}
\description{
Save plots to PNG + SVG files
}
