% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hsci.R
\name{theme_hsci}
\alias{theme_hsci}
\title{HSCI ggplot theme}
\usage{
theme_hsci(base_size = 12, base_family = "sans")
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
HSCI ggplot theme
}
\seealso{
\code{\link{ggplot}}
}
