% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hsci.R
\name{theme_hsci_continuous}
\alias{theme_hsci_continuous}
\title{HSCI ggplot theme with viridis continuous palette}
\usage{
theme_hsci_continuous(
  base_size = 12,
  base_family = "Helvetica",
  palette = "viridis"
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{palette}{viridis palette to use. Four options are available: "magma" (or "A"), "inferno" (or "B"), "plasma" (or "C"), "viridis" (or "D", the default option) and "cividis" (or "E").}
}
\description{
HSCI ggplot theme with viridis continuous palette
}
\seealso{
\code{\link{theme_hsci}}
}
