% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hsci.R
\name{theme_hsci_continuous_viridis_grayscale}
\alias{theme_hsci_continuous_viridis_grayscale}
\title{HSCI ggplot theme with grayscale viridis continuous palette}
\usage{
theme_hsci_continuous_viridis_grayscale(
  base_size = 12,
  base_family = "Helvetica",
  palette = "viridis"
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{palette}{virids palette to use}
}
\description{
HSCI ggplot theme with grayscale viridis continuous palette
}
\seealso{
\code{\link{theme_hsci}}
}
