% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hsci.R
\name{theme_hsci_discrete}
\alias{theme_hsci_discrete}
\title{HSCI ggplot theme with a discrete coloropt palette.}
\usage{
theme_hsci_discrete(
  base_size = 12,
  base_family = "Helvetica",
  palette = "normal"
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{palette}{coloropt palette from normal, bright, dark, fancy and tarnish}
}
\description{
HSCI ggplot theme with a discrete coloropt palette.
}
\seealso{
\code{\link{theme_hsci}}
}
