% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hsci.R
\name{theme_hsci_discrete_grayscale}
\alias{theme_hsci_discrete_grayscale}
\title{HSCI ggplot theme with a grayscale coloropt discrete palette.}
\usage{
theme_hsci_discrete_grayscale(
  base_size = 12,
  base_family = "Helvetica",
  palette = "normal"
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{palette}{coloropt palette from normal, bright, dark, fancy and tarnish}
}
\description{
HSCI ggplot theme with a grayscale coloropt discrete palette.
}
