% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hsci.R
\name{theme_hsci_discrete_viridis}
\alias{theme_hsci_discrete_viridis}
\title{HSCI ggplot theme with viridis discrete palette}
\usage{
theme_hsci_discrete_viridis(
  base_size = 12,
  base_family = "Helvetica",
  palette = "viridis"
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{palette}{virids palette to use}
}
\description{
HSCI ggplot theme with viridis discrete palette
}
\seealso{
\code{\link{theme_hsci}}
}
